add new allure report
