+++ src/test/java/RegistrationTest.java
@@ -2,6 +2,7 @@ import com.codeborne.selenide.WebDriverRunner;
 import dataGenerator.DefaultUserData;
 import deleteData.DeleteUser;
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import pageDescription.RegistrationPage;
@@ -20,21 +21,16 @@ public class RegistrationTest {
         open(registrationPageUrl);
         registrationPage = new RegistrationPage();
     }
+    @After
+    public void deleteUser() {
+        DeleteUser deleteUser = new DeleteUser();
+        deleteUser.deleteUser();
+    }
     @Test
     @DisplayName("Registration new user and check URL and login button")
     public void RegistrationUserTest() {
         registrationPage.setRegistrationFields(DefaultUserData.DEFAULT_USER_NAME,DefaultUserData.DEFAULT_USER_EMAIL,DefaultUserData.DEFAULT_USER_PASSWORD);
         String currentUrl = WebDriverRunner.getWebDriver().getCurrentUrl();
         assertEquals("Текущий URL не совпадает с URL страницы входа",loginUrl, currentUrl);
-        DeleteUser deleteUser = new DeleteUser();
-        deleteUser.deleteUser();
-    }
-
-    @Test
-    @DisplayName("Registration new user with wrong password and check error message")
-    public void passwordErrorTest() {
-        registrationPage.setPassword(DefaultUserData.ERROR_USER_PASSWORD);
-        registrationPage.clickRegistration();
-        registrationPage.checkVisibleErrorPasswordMessage();
     }
 }
